requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
attrs==23.2.0
cbor2==5.7.1
certifi==2025.11.12
charset-normalizer==3.4.4
comet_ml==3.54.2
configobj==5.0.9
contourpy==1.3.3
cycler==0.12.1
diffusion==6.12.1
diffusion-core==0.0.71
dulwich==0.24.10
einops==0.8.1
everett==3.1.0
filelock==3.20.0
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.12.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
networkx==3.6
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pillow==12.0.0
propcache==0.4.1
psutil==7.1.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-box==6.1.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
semantic-version==2.10.0
sentry-sdk==2.47.0
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
stringcase==1.2.0
structlog==21.5.0
sympy==1.14.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.0
wrapt==2.0.1
wurlitzer==3.1.1
yarl==1.22.0
